\name{format.timeDate}
\alias{format.timeDate}
\alias{format.timeSpan}
\title{
  Lists of Sample Formats
}
\description{
The \code{format.timeDate}  and \code{format.timeSpan} data sets are
lists with sample formats for the \code{timeDate}  and
\code{timeSpan}  classes, respectively.  
\itemize{
\item The names on the lists are the sample
strings for each format.
\item The corresponding elements are
lists whose component is the output format that generates the 
sample string, and whose input component is the
input format that reads in the sample string. 
}

See \code{class.timeDate} for information on
input and output string formats.
}
\seealso{
  \code{\linkS4class{timeDate}}  class,
  \code{\linkS4class{timeSpan}}  class.
}
\keyword{datasets}
